--- src/main/java/juja/microservices/teams/exceptions/TeamsException.java
@@ -1,7 +1,0 @@
-package juja.microservices.teams.exceptions;
-
-public class TeamsException extends RuntimeException {
-    public TeamsException(String message) {
-        super(message);
-    }
-}
