--- src/main/java/juja/microservices/teams/exceptions/UserAlreadyInTeamException.java
@@ -1,8 +1,0 @@
-package juja.microservices.teams.exceptions;
-
-/**
- * Created by Andrii Sidun
- */
-public class UserAlreadyInTeamException extends RuntimeException {
-    public UserAlreadyInTeamException(String message) {super(message);}
-}
