--- src/main/java/juja/microservices/teams/exceptions/UserInSeveralTeamsException.java
@@ -1,7 +1,0 @@
-package juja.microservices.teams.exceptions;
-
-public class UserInSeveralTeamsException extends TeamsException {
-    public UserInSeveralTeamsException(String message) {
-        super(message);
-    }
-}
