--- src/main/java/juja/microservices/teams/exceptions/UserNotInTeamException.java
@@ -1,7 +1,0 @@
-package juja.microservices.teams.exceptions;
-
-public class UserNotInTeamException extends TeamsException {
-    public UserNotInTeamException(String message) {
-        super(message);
-    }
-}
