--- src/main/java/juja/microservices/teams/exceptions/UserNotKeeperException.java
@@ -1,8 +1,0 @@
-package juja.microservices.teams.exceptions;
-
-/**
- * Created by Andrii Sidun
- */
-public class UserNotKeeperException extends RuntimeException {
-    public UserNotKeeperException(String message) {super(message);}
-}
