+++ src/main/java/ua/com/juja/microservices/teams/entity/UserUuidRequest.java
@@ -1,0 +1,17 @@
+package ua.com.juja.microservices.teams.entity;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import org.hibernate.validator.constraints.NotEmpty;
+
+@Getter
+public class UserUuidRequest {
+    @NotEmpty
+    private String uuid;
+
+    @JsonCreator
+    public UserUuidRequest(@JsonProperty("from") String uuid) {
+        this.uuid = uuid;
+    }
+}
