+++ src/main/java/ua/com/juja/microservices/teams/exceptions/TeamsException.java
@@ -1,0 +1,7 @@
+package ua.com.juja.microservices.teams.exceptions;
+
+public class TeamsException extends RuntimeException {
+    public TeamsException(String message) {
+        super(message);
+    }
+}
