+++ src/main/java/ua/com/juja/microservices/teams/exceptions/UserAlreadyInTeamException.java
@@ -1,0 +1,8 @@
+package ua.com.juja.microservices.teams.exceptions;
+
+/**
+ * Created by Andrii Sidun
+ */
+public class UserAlreadyInTeamException extends RuntimeException {
+    public UserAlreadyInTeamException(String message) {super(message);}
+}
