+++ src/main/java/ua/com/juja/microservices/teams/exceptions/UserInSeveralTeamsException.java
@@ -1,0 +1,7 @@
+package ua.com.juja.microservices.teams.exceptions;
+
+public class UserInSeveralTeamsException extends TeamsException {
+    public UserInSeveralTeamsException(String message) {
+        super(message);
+    }
+}
