+++ src/main/java/ua/com/juja/microservices/teams/exceptions/UserNotInTeamException.java
@@ -1,0 +1,7 @@
+package ua.com.juja.microservices.teams.exceptions;
+
+public class UserNotInTeamException extends TeamsException {
+    public UserNotInTeamException(String message) {
+        super(message);
+    }
+}
