+++ src/main/java/ua/com/juja/microservices/teams/exceptions/UserNotKeeperException.java
@@ -1,0 +1,8 @@
+package ua.com.juja.microservices.teams.exceptions;
+
+/**
+ * Created by Andrii Sidun
+ */
+public class UserNotKeeperException extends RuntimeException {
+    public UserNotKeeperException(String message) {super(message);}
+}
