--- src/test/java/juja/microservices/integration/TeamControllerIntegrationTest.java
@@ -1,62 +1,0 @@
-package juja.microservices.integration;
-
-import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-
-import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-/**
- * @author Ivan Shapovalov
- * @author Andrii Sidun
- */
-@RunWith(SpringRunner.class)
-public class TeamControllerIntegrationTest extends BaseIntegrationTest {
-    private static final String TEAMS_DEACTIVATE_URL = "/v1/teams/users/";
-
-    private MockMvc mockMvc;
-
-    @Rule
-    public final ExpectedException exception = ExpectedException.none();
-
-    @Before
-    public void setup() {
-        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
-    }
-
-    @Test
-    @UsingDataSet(locations = "/datasets/deactivateTeam_dataSet.json")
-    public void test_deactivateTeamIfUserInTeamExecutedCorrectly() throws Exception {
-        final String uuid = "user-in-one-team";
-        mockMvc.perform(put(TEAMS_DEACTIVATE_URL+uuid))
-                .andExpect(content().contentType(APPLICATION_JSON_UTF8))
-                .andExpect(status().isOk());
-    }
-
-    @Test
-    @UsingDataSet(locations = "/datasets/deactivateTeam_dataSet.json")
-    public void test_deactivateTeamIfUserNotInTeamExecutedCorrectly() throws Exception {
-        final String uuid = "user-not-in-team";
-        mockMvc.perform(put(TEAMS_DEACTIVATE_URL+uuid))
-                .andExpect(content().contentType(APPLICATION_JSON_UTF8))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    @UsingDataSet(locations = "/datasets/deactivateTeam_dataSet.json")
-    public void test_deactivateTeamIfUserInSeveralTeamsExecutedCorrectly() throws Exception {
-        final String uuid = "user-in-several-teams";
-        mockMvc.perform(put(TEAMS_DEACTIVATE_URL+uuid))
-                .andExpect(content().contentType(APPLICATION_JSON_UTF8))
-                .andExpect(status().isBadRequest());
-    }
-}
