--- src/test/resources/acceptance/response/responseDeactivateTeamIfUserInSeveralTeamsThrowsExceptions.json
@@ -1,8 +1,0 @@
-{
-  "httpStatus": 400,
-  "internalErrorCode": "TMF-F2-D2",
-  "clientMessage": "You cannot deactivate team if user in several teams",
-  "developerMessage": "The reason of the exception is that user in several teams",
-  "exceptionMessage": "User with uuid 'user-in-several-teams' is in several teams now",
-  "detailErrors": []
-}
