+++ src/test/resources/acceptance/response/responseGetDeactivateTeamIfUserInTeamExecutedCorrectly.json
