+++ src/test/resources/acceptance/response/responseGetDeactivateTeamIfUserNotInTeamThrowsExeption.json
